requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.23
pandas==2.1.3
pydantic==2.5.0
python-dotenv==1.0.0
bcrypt==4.1.2
pandas_ta
SQLAlchemy
pymongo==4.6.0
motor==3.3.2